Multi-output regression
Allows basic multi-output regression.  Splits according to
the average mse across the outputs.  Each leaf contains a
prediction for each output.  Though an improvement to the
equivalent use in RF (e.g. the scikit learn version), this
can still have odd behaviour when the outputs are non-Gaussian
distributed as the mean of the tree predictions in multiple
dimensions can be an innappropriate predictor.